backend/requirements.txt
sqlalchemy==2.0.28
alembic==1.13.1
asyncpg==0.29.0
python-dotenv==1.0.1 